M9/ex1/requirements.txt
pydantic2
flake8
mypy